feat(dashboard): copy buttons on every code step in /setup onboarding
Caught during E2E: only the multi-line code block on /setup had a
copy button (via the empty-state CopyButton). The Install / Run
shell-command steps showed `pip install awaithumans` and `python
refund.py` as plain monospace with no way to copy short of
hand-selection, which fights operators who are mid-flow.

## Change

Extracted the existing copy button into a shared
`components/copy-button.tsx` so the empty-state and the setup
onboarding render the same control:

  - `<CopyButton code={...} />` — flips to "copied" affirmation for
    1.5s after a successful clipboard write, silent fallback when
    the API is unavailable (insecure context, etc.)
  - Shell-empty-state now re-exports the shared button instead of
    its own local copy
  - Setup page's `CodeBlock` accepts `code: string` (was: ReactNode
    children), renders the copy button absolute-positioned in the
    top-right of the block

## Visibility rule

Inline `pip install` / `python refund.py` etc. are short shell
commands — `opacity-0 group-hover:opacity-100` on those keeps the
prompt clean at rest, button surfaces on hover.

Multi-line code blocks (the full agent example) are explicit
"reach for this" surfaces — copy button stays visible at all
times since the user is definitely going to want it.

## Files
  - new `components/copy-button.tsx` — shared control
  - `components/shell-empty-state.tsx` — drop local copy, import shared
  - `app/setup/page.tsx` — wire copy button into all three CodeBlocks
    (install / save / run)
