+++ packages/dashboard/components/shell-empty-state.tsx
@@ -11,9 +11,9 @@
  * screen" for a lot of people — make it feel intentional.
  */
 
-import { Check, Copy } from "lucide-react";
 import { useState } from "react";
 
+import { CopyButton } from "@/components/copy-button";
 import { Eyebrow } from "@/components/eyebrow";
 import { cn } from "@/lib/utils";
 
@@ -150,42 +150,3 @@ function CodeSnippet({
 	);
 }
 
-function CopyButton({ code }: { code: string }) {
-	const [copied, setCopied] = useState(false);
-
-	const copy = async () => {
-		try {
-			await navigator.clipboard.writeText(code);
-			setCopied(true);
-			setTimeout(() => setCopied(false), 1500);
-		} catch {
-			// Clipboard API blocked (insecure context); fall through silently.
-		}
-	};
-
-	return (
-		<button
-			type="button"
-			onClick={copy}
-			className={cn(
-				"inline-flex items-center gap-1.5 text-[11px] font-mono px-2 py-1 rounded transition-colors",
-				copied
-					? "text-brand bg-brand/5"
-					: "text-white/40 hover:text-white hover:bg-white/5",
-			)}
-			aria-label={copied ? "Copied" : "Copy code"}
-		>
-			{copied ? (
-				<>
-					<Check size={11} />
-					copied
-				</>
-			) : (
-				<>
-					<Copy size={11} />
-					copy
-				</>
-			)}
-		</button>
-	);
-}
